ft-#22: implement update passcode test util
